+++ tests/test_donut_viz_pipeline_cwfs_sensors.py
@@ -110,7 +110,9 @@ class TestDonutVizPipeline(TestCase):
             agg_zern_raw["detector"],
             sorted([det for det in self.extraDetectorNames for _ in range(2)]),
         )
-        self.assertCountEqual(agg_zern_raw.meta.keys(), self.meta_keys)
+        self.assertCountEqual(
+            agg_zern_raw.meta.keys(), self.meta_keys + ["estimatorInfo"]
+        )
 
     def testAggregateDonuts(self):
         donut_table_list = list(
@@ -168,7 +170,9 @@ class TestDonutVizPipeline(TestCase):
         self.assertEqual(len(raw_visit_table_list), 1)
         self.assertEqual(raw_visit_table_list[0].dataId["visit"], 4021123106000)
         raw_visit_table = self.butler.get(raw_visit_table_list[0])
-        self.assertCountEqual(raw_visit_table.meta.keys(), self.meta_keys)
+        self.assertCountEqual(
+            raw_visit_table.meta.keys(), self.meta_keys + ["estimatorInfo"]
+        )
         raw_zern_table = self.butler.get(
             "aggregateZernikesRaw",
             dataId=raw_visit_table_list[0].dataId,